+++ CHANGELOG.md
@@ -6,4 +6,5 @@
 - [tab] feature: optimize the `defualt config file` for this module. (Now it provide a better color schema in the default example)
 - [core] feature: improvements to `language` system.
   - feature: increase the number of supported languages from `30` to all `128` languages. (even if the language is not built-in, users can now still use it.)
+  - feature: the `language files` now will be reloaded by `/fuji reload` command, regardless of whether the `language` module is enabled or not.
   - fix: when `language` module is enabled, and a player joins with an `unsupported language` the first time, the console should not be spammed with exceptions.